+++ R/geom_flame.R
@@ -10,10 +10,6 @@
 #' of this being that one may add additional information to the figure as geom
 #' layers to ggplot2 graphs as may be necessary.
 #'
-#' @import ggplot2
-#' @importFrom ggplot2 ggproto
-#' @inheritParams ggplot2::layer
-#'
 #' @seealso \code{\link{event_line}} for a non-ggplot2 based flame function.
 #'
 #' @section Aesthetics:
@@ -30,6 +26,33 @@
 #'   \item \code{linetype}
 #' }
 #'
+#' @param mapping Set of aesthetic mappings created by \code{aes()} or \code{aes_()}. If
+#' specified and inherit.aes = TRUE (the default), it is combined with the
+#' default mapping at the top level of the plot. You must supply mapping if
+#' there is no plot mapping.
+#' @param data The data to be displayed in this layer. There are three options:
+#'
+#' If NULL, the default, the data is inherited from the plot data as specified
+#' in the call to \code{ggplot()}.
+#'
+#' A data.frame, or other object, will override the plot data. All objects will
+#' be fortified to produce a data frame. See \code{fortify()} for which variables will
+#' be created.
+#'
+#' A function will be called with a single argument, the plot data. The return
+#' value must be a \code{data.frame}, and will be used as the layer data.
+#' @param stat The statistical transformation to use on the data for this layer,
+#' as a string.
+#' @param position Position adjustment, either as a string, or the result of a call
+#' to a position adjustment function.
+#' @param show.legend Logical. Should this layer be included in the legends? \code{NA},
+#' the default, includes if any aesthetics are mapped. \code{FALSE} never includes, and
+#' \code{TRUE} always includes. It can also be a named logical vector to finely select
+#' the aesthetics to display.
+#' @param inherit.aes If \code{FALSE}, overrides the default aesthetics, rather
+#' than combining with them. This is most useful for helper functions that define
+#' both data and aesthetics and shouldn't inherit behaviour from the default plot
+#' specification, e.g. \code{borders()}.
 #' @param na.rm If \code{FALSE} (the default), removes missing values with
 #'    a warning. If \code{TRUE} silently removes missing values.
 #' @param ... other arguments passed on to \code{\link{layer}}. These are
@@ -46,8 +69,8 @@
 #' @export
 #'
 #' @examples
-#' t_dat <- make_whole(sst_WA)
-#' res <- detect(t_dat, climatology_start = 1983, climatology_end = 2012)
+#' ts_dat <- make_whole(sst_WA)
+#' res <- detect(ts_dat, climatology_start = 1983, climatology_end = 2012)
 #' mhw <- res$clim
 #' mhw <- mhw[10580:10690,]
 #'
@@ -85,10 +108,10 @@ GeomFlame <- ggplot2::ggproto("GeomFlame", ggplot2::Geom,
 
                               required_aes = c("x", "y", "y2"),
 
-                              default_aes = aes(colour = NA, fill = "salmon",
+                              default_aes = ggplot2::aes(colour = NA, fill = "salmon",
                                                 size = 0.5, linetype = 1, alpha = NA),
 
-                              draw_key = draw_key_polygon,
+                              draw_key = ggplot2::draw_key_polygon,
 
                               draw_group = function(data, panel_scales, coord, na.rm = FALSE) {
                                 if (na.rm) data <- data[stats::complete.cases(data[c("x", "y", "y2")]), ]