+++ NEWS.md
@@ -3,9 +3,13 @@ output: pdf_document
 ---
 
 # VERSION CHANGES
-## Changes in version 0.16.0 (30 December 2017)
-* The make_whole() function now handles any arbitrary column names, which are specified 
-via arguments to the function.
+## Changes in version 0.16.0 (6 January 2018)
+* The make_whole(), detect() and exceedance() functions now handle any arbitrary
+column names, which are specified via arguments to the function.
+* Remove dependence on magrittr.
+* More consistent use of tidyverse packages throughout.
+* Updates to examples.
+* Miscellaneous other improvements troughout the function internals.
 
 ## Changes in version 0.15.9 (19 June 2017)
 * Warnings given by exceedance() when exceedances begin/end on the first/last day of the